Pruebas unitarias autowired and delete (Andrés)
