+++ src/test/java/com/spring/proyecto2/Pruebas_JPA_BorrarContacto.java
@@ -1,5 +1,7 @@
 package com.spring.proyecto2;
 
+import javax.transaction.Transactional;
+
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -28,5 +30,21 @@ public class Pruebas_JPA_BorrarContacto extends TestCase{
 		else {
 			assertEquals((cantidad-1), repository.count());
 		}
+	}
+	
+	@Test
+	@Transactional
+	public void testDeleteTable() {
+		repository.deleteAll();
+		assertTrue(repository.findAll().size() == 0);
+	}
+	
+	@Test
+	@Transactional
+	public void testDeleteId() {
+		int init = repository.findAll().size();
+		if(repository.findById(2)!=null)
+			repository.deleteById(2);
+		assertTrue((init-1) == repository.findAll().size());
 	}
 }
